add main

+++ test/transpiler/pheno_pkg/crop2ml/algo/pyx/ShootNumber.pyx
@@ -4,7 +4,7 @@ cdef int emergedLeaves, shoots, i
 oldCanopyShootNumber = canopyShootNumber
 emergedLeaves = int(max(1.0, ceil(leafNumber - 1)))
 shoots = fibonacci_(emergedLeaves)
-canopyShootNumber = min(float(shoots * sowingDensity), targetFertileShoot)
+canopyShootNumber = min(shoots * sowingDensity, targetFertileShoot)
 averageShootNumberPerPlant = canopyShootNumber / sowingDensity       
 if (canopyShootNumber != oldCanopyShootNumber):
     tilleringProfile.append(canopyShootNumber - oldCanopyShootNumber)
